+++ src/main/java/com/sda/course/project/restaurant/service/OrderService.java
@@ -5,9 +5,9 @@ import com.sda.course.project.restaurant.repository.OrderRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import javax.validation.constraints.NotNull;
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class OrderService {
@@ -18,8 +18,8 @@ public class OrderService {
         return orderRepository.findAll();
     }
 
-    public OrderEntity getOrderById(Integer id) {
-        return orderRepository.findById(id).orElse(null);
+    public Optional<OrderEntity> getOrderById(Integer id) {
+        return orderRepository.findById(id);
     }
 
     public List<OrderEntity> getOrdersByPerson(Integer id) {
@@ -32,20 +32,34 @@ public class OrderService {
 
     public List<OrderEntity> getTopTenUsersByOrder() {
         return orderRepository.findTop10ByOrderByTotalPriceDesc();
-//        orderRepository.findAll();
-
     }
 
-    public OrderEntity addOrder(@NotNull OrderEntity order) {
-        order.setDate(new Date());
-        return orderRepository.save(order);
+    public OrderEntity addOrder(OrderEntity newOrder) {
+        newOrder.setDate(new Date());
+        return orderRepository.save(newOrder);
     }
 
-    public void deleteOrderById(Integer id) {
-        orderRepository.deleteById(id);
+    public Optional<OrderEntity> deleteOrderById(Integer id) {
+        Optional<OrderEntity> orderToDelete = orderRepository.findById(id);
+
+        if(orderToDelete.isPresent()){
+            orderRepository.deleteById(id);
+        }
+
+        return orderToDelete;
     }
 
-    public void updateOrder(OrderEntity updatedOrder) {
-        orderRepository.save(updatedOrder);
+    public OrderEntity updateOrder(Integer id, OrderEntity updatedOrder) {
+        return orderRepository.findById(id)
+                .map(o -> {
+                    o.setDate(updatedOrder.getDate());
+                    o.setPerson(updatedOrder.getPerson());
+                    o.setTotalPrice(updatedOrder.getTotalPrice());
+
+                    return orderRepository.save(o);
+                }).orElseGet(() -> {
+                    updatedOrder.setOrderId(id);
+                    return orderRepository.save(updatedOrder);
+                });
     }
 }
